Move run-remote-query to remote queries area
