+++ extensions/ql-vscode/src/remote-queries/run-remote-query.ts
@@ -3,13 +3,13 @@ import * as path from 'path';
 import * as yaml from 'js-yaml';
 import * as fs from 'fs-extra';
 import * as tmp from 'tmp-promise';
-import { askForLanguage, findLanguage, getOnDiskWorkspaceFolders, showAndLogErrorMessage, showAndLogInformationMessage, showAndLogWarningMessage, showInformationMessageWithAction } from './helpers';
-import { Credentials } from './authentication';
-import * as cli from './cli';
-import { logger } from './logging';
-import { getRemoteControllerRepo, getRemoteRepositoryLists, setRemoteControllerRepo } from './config';
-import { tmpDir } from './run-queries';
-import { ProgressCallback, UserCancellationException } from './commandRunner';
+import { askForLanguage, findLanguage, getOnDiskWorkspaceFolders, showAndLogErrorMessage, showAndLogInformationMessage, showAndLogWarningMessage, showInformationMessageWithAction } from '../helpers';
+import { Credentials } from '../authentication';
+import * as cli from '../cli';
+import { logger } from '../logging';
+import { getRemoteControllerRepo, getRemoteRepositoryLists, setRemoteControllerRepo } from '../config';
+import { tmpDir } from '../run-queries';
+import { ProgressCallback, UserCancellationException } from '../commandRunner';
 import { OctokitResponse } from '@octokit/types/dist-types';
 import * as unzipper from 'unzipper';
