+++ extensions/ql-vscode/src/vscode-tests/cli-integration/run-remote-query.test.ts
@@ -7,7 +7,7 @@ import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as yaml from 'js-yaml';
 
-import { QlPack, runRemoteQuery } from '../../run-remote-query';
+import { QlPack, runRemoteQuery } from '../../remote-queries/run-remote-query';
 import { Credentials } from '../../authentication';
 import { CliVersionConstraint, CodeQLCliServer } from '../../cli';
 import { CodeQLExtensionInterface } from '../../extension';
